+++ elm_bk/src/main/java/com/tju/elm_bk/mapper/BusinessMapper.java
@@ -15,23 +15,23 @@ import org.apache.ibatis.annotations.*;
 @Mapper
 public interface BusinessMapper {
 
-    BusinessVO getBusinessById(Integer businessId);
+    BusinessVO getBusinessById(Long businessId);
 
     @Select("SELECT * FROM business WHERE id = #{id}")
     BusinessPermissionVO getBusinessPermissionById(Long businessId);
 
     // 更新商户信息
-    int updateBusiness(@Param("id") Integer id, @Param("updateDto") BusinessUpdateDTO updateDto);
+    int updateBusiness(@Param("id") Long id, @Param("updateDto") BusinessUpdateDTO updateDto);
 
     // 更新商户所有者信息
-    int updateBusinessOwner(@Param("id") Integer id, @Param("updateDto") BusinessUpdateDTO updateDto);
+    int updateBusinessOwner(@Param("id") Long id, @Param("updateDto") BusinessUpdateDTO updateDto);
 
     // 嵌套查询方法
-    List<Authority> selectAuthoritiesByUserId(@Param("userId") Integer userId);
+    List<Authority> selectAuthoritiesByUserId(@Param("userId") Long userId);
 
     // 逻辑删除商户并返回删除前的信息
     @Update("UPDATE business SET is_deleted = 1 WHERE id = #{id} AND is_deleted = 0")
-    int deleteBusiness(@Param("id") Integer id);
+    int deleteBusiness(@Param("id") Long id);
 
     int insertBusiness(@Param("businessDto") BusinessDTO businessDto);
