+++ elm_bk/src/main/java/com/tju/elm_bk/service/BusinessService.java
@@ -7,10 +7,10 @@ import com.tju.elm_bk.vo.BusinessVO;
 import java.util.List;
 
 public interface BusinessService {
-    BusinessVO getBusinessById(Integer id);
-    BusinessVO updateBusiness(Integer id, BusinessUpdateDTO updateDto);
-    BusinessVO deleteBusiness(Integer id);
-    BusinessVO patchBusiness(Integer id, BusinessUpdateDTO updateDto);
+    BusinessVO getBusinessById(Long id);
+    BusinessVO updateBusiness(Long id, BusinessUpdateDTO updateDto);
+    BusinessVO deleteBusiness(Long id);
+    BusinessVO patchBusiness(Long id, BusinessUpdateDTO updateDto);
     List<BusinessVO> getBusinesses();
     BusinessVO addBusiness(BusinessDTO businessDto);
 }
